# crates/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Squire Local API",
    "description": "The LAN HTTP wire contract for the Squire household reward system — emitted from the Rust types (ADR SQUIRE-A-0009). IDs are int64 on the wire (a 'fits in i64' invariant).",
    "license": {
      "name": "Proprietary",
      "identifier": "LicenseRef-Proprietary"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/members/{id}/tincture": {
      "post": {
        "tags": [
          "knight"
        ],
        "summary": "`POST /admin/members/{id}/tincture` ([`RequireKnight`]) — a Knight sets a member's colour.",
        "operationId": "set_member_tincture",
        "parameters": [
          {
            "name": "X-Household",
            "in": "header",
            "description": "Opaque household handle routing the request to its tenant",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Member user id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTinctureReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Tincture saved"
          },
          "400": {
            "description": "Not one of the known tinctures"
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "403": {
            "description": "Authenticated but not a Knight"
          },
          "404": {
            "description": "No such member"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/me/tincture": {
      "post": {
        "tags": [
          "squire"
        ],
        "summary": "`POST /me/tincture` ([`RequireSquire`]) — the signed-in squire picks their own colour.",
        "operationId": "set_my_tincture",
        "parameters": [
          {
            "name": "X-Household",
            "in": "header",
            "description": "Opaque household handle routing the request to its tenant",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTinctureReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Tincture saved"
          },
          "400": {
            "description": "Not one of the known tinctures"
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "403": {
            "description": "Not a Squire"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SetTinctureReq": {
        "type": "object",
        "description": "`POST /me/tincture` / `POST /admin/members/{id}/tincture` body.",
        "required": [
          "tincture"
        ],
        "properties": {
          "tincture": {
            "type": "string",
            "description": "One of `gules`, `azure`, `vert`, `purpure`, `tenne`, `sable`."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Tenant-scoped bearer token (proves household, user, role)."
      }
    }
  },
...
}